refactor: take one tool's implementation out of two oversized classes
DebugTools exposes six tools and a quarter of the file was the reflection
machinery behind one of them, list_event_handlers. TinkerTools' own job is
running code safely, and most of it was laying out the transcript afterwards.
Both are now collaborators: support\EventHandlers and support\Transcript.

  DebugTools   505 -> 363
  TinkerTools  322 -> 206

Also closes the outputSchema hole in the Presenter. It dropped
structuredContent unconditionally, which is right while no tool declares an
output schema, and silently wrong the day one does: the server would advertise
a schema and never send the payload the spec ties to it. Dormant today, pinned
by a test, and the failure it prevents would not have announced itself.

All four smoke profiles unchanged.
