+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tool annotations describe what each tool actually does. `create_entry`, `duplicate_entry`, `copy_entry_to_site` and `create_nested_entry` advertised `destructiveHint: true` while being purely additive, and `run_query` advertised it while being SELECT-only under `SqlReadGuard`; `move_nested_entry` gains `idempotentHint: true`, since moving a block to position N twice leaves the same state, and `reload_mcp` gains annotations at all instead of inheriting the spec's conservative defaults for a call that changes no data. Every tool now also declares `openWorldHint: false`, which is accurate for a server that only ever touches one Craft install. None of this changes what a token may call: access is gated by the `dangerous` flag in our own metadata, and these hints only tell a client when to ask for confirmation. Prompting on read-only and additive calls trains people to click through the prompts that matter
 
 ### Fixed
+- A tool that declares an `outputSchema` now receives the matching `structuredContent`. The duplicate is still dropped for every tool that declares none, which is all of them today, so nothing changes on the wire; without the guard, the first tool to declare a schema would have advertised it and silently never honoured it
 - Every tool, prompt and resource body is now just its own work. The wrapper each one opened to convert exceptions and run the freshness probe is gone: both jobs moved to decorators that cover every call, so 78 closures and three helper classes came out with no change to what any tool returns. An architecture test keeps them from creeping back
 - An error thrown outside a tool body keeps its message. Anything that failed during argument preparation, result formatting, or inside the output layer itself reached the SDK's generic catch and came back as the fixed string `Error while executing tool`, with the real cause discarded and nothing for an agent to act on. That is not hypothetical: the fatal that broke every parameterless tool on this branch surfaced exactly that way. A caller mistake keeps its own wording, because "Missing required argument `id`" is more use than the same text behind a file and line number
 - The project-config freshness probe can finally notify. It only tells subscribers when handed a request context, and 4 of 62 call sites passed one, so it ran on every tool call and could reach nobody. It now runs once per tool call with the context the SDK already injects
