+++ src/tools/DebugTools.php
@@ -4,21 +4,17 @@ declare(strict_types=1);
 
 namespace stimmt\craft\Mcp\tools;
 
-use Closure;
 use Craft;
-use Generator;
 use Mcp\Capability\Attribute\McpTool;
 use Mcp\Exception\ToolCallException;
 use Mcp\Schema\ToolAnnotations;
 use Mcp\Server\RequestContext;
-use ReflectionClass;
-use ReflectionFunction;
 use stimmt\craft\Mcp\attributes\McpToolMeta;
 use stimmt\craft\Mcp\enums\ToolCategory;
+use stimmt\craft\Mcp\support\EventHandlers;
 use stimmt\craft\Mcp\support\FileHelper;
 use stimmt\craft\Mcp\support\SqlReadGuard;
 use Throwable;
-use yii\base\Event;
 
 /**
  * Debugging tools for Craft CMS.
@@ -348,8 +344,9 @@ class DebugTools {
     )]
     #[McpToolMeta(category: ToolCategory::DEBUGGING)]
     public function listEventHandlers(?string $filter = null, ?RequestContext $context = null): array {
-        $handlers = $this->getApplicationEvents($filter);
-        $classEvents = $this->getClassEvents($filter);
+        $events = new EventHandlers();
+        $handlers = $events->applicationEvents($filter);
+        $classEvents = $events->classEvents($filter);
 
         return [
             'applicationEvents' => [
@@ -362,144 +359,5 @@ class DebugTools {
             ],
             'hint' => 'Use filter parameter to search by event or class name',
         ];
-    }
-
-    /**
-     * Get application-level event handlers.
-     */
-    private function getApplicationEvents(?string $filter): array {
-        try {
-            $reflection = new ReflectionClass(Craft::$app);
-            $eventsProperty = $reflection->getProperty('_events');
-            $events = $eventsProperty->getValue(Craft::$app) ?? [];
-        } catch (Throwable) {
-            return [];
-        }
-
-        $handlers = [];
-        foreach ($events as $eventName => $eventHandlers) {
-            if ($filter !== null && stripos((string) $eventName, $filter) === false) {
-                continue;
-            }
-
-            $handlers[$eventName] = [
-                'count' => count($eventHandlers),
-                'handlers' => $this->describeHandlers($eventHandlers),
-            ];
-        }
-
-        return $handlers;
-    }
-
-    /**
-     * Get class-level event handlers (Events::on()).
-     */
-    private function getClassEvents(?string $filter): array {
-        try {
-            $eventReflection = new ReflectionClass(Event::class);
-            $classEventsProperty = $eventReflection->getStaticPropertyValue('_events');
-        } catch (Throwable) {
-            return [];
-        }
-
-        $classEvents = [];
-        foreach ($this->flattenClassEvents($classEventsProperty, $filter) as $event) {
-            $key = "{$event['class']}::{$event['event']}";
-            $classEvents[$key] = $event;
-        }
-
-        return $classEvents;
-    }
-
-    /**
-     * Flatten nested class events structure.
-     *
-     * Yii stores class-level handlers as `Event::$_events[$eventName][$className][]`,
-     * so the outer key is the event name and the inner key is the class name.
-     *
-     * @return Generator<array{class: string, event: string, count: int, handlers: array}>
-     */
-    private function flattenClassEvents(array $classEventsProperty, ?string $filter): Generator {
-        $flattened = array_merge(...array_map(
-            fn (string $eventName, array $classHandlers) => $this->extractClassEvents($eventName, $classHandlers, $filter),
-            array_keys($classEventsProperty),
-            array_values($classEventsProperty),
-        ));
-
-        yield from $flattened;
-    }
-
-    /**
-     * Extract the per-class handlers registered for a single event name.
-     *
-     * @param array<string, array> $classHandlers keyed by class name
-     * @return array<array{class: string, event: string, count: int, handlers: array}>
-     */
-    private function extractClassEvents(string $eventName, array $classHandlers, ?string $filter): array {
-        return array_filter(
-            array_map(
-                fn (string $className, array $eventHandlerList) => $this->matchesFilter($className, $eventName, $filter)
-                    ? [
-                        'class' => $className,
-                        'event' => $eventName,
-                        'count' => count($eventHandlerList),
-                        'handlers' => $this->describeHandlers($eventHandlerList),
-                    ]
-                    : null,
-                array_keys($classHandlers),
-                array_values($classHandlers),
-            ),
-        );
-    }
-
-    /**
-     * Check if class or event name matches filter.
-     */
-    private function matchesFilter(string $className, string $eventName, ?string $filter): bool {
-        if ($filter === null) {
-            return true;
-        }
-
-        return stripos($eventName, $filter) !== false || stripos($className, $filter) !== false;
-    }
-
-    /**
-     * Describe a list of event handlers.
-     */
-    private function describeHandlers(array $handlers): array {
-        return array_map(
-            fn (array $handler) => $this->describeCallback($handler[0]),
-            $handlers,
-        );
-    }
-
-    /**
-     * Describe a callback for human readability.
-     */
-    private function describeCallback(mixed $callback): string {
-        if (is_string($callback)) {
-            return $callback;
-        }
-
-        if (is_array($callback)) {
-            $class = is_object($callback[0]) ? $callback[0]::class : $callback[0];
-            $method = $callback[1];
-
-            return "{$class}::{$method}()";
-        }
-
-        if ($callback instanceof Closure) {
-            $reflection = new ReflectionFunction($callback);
-            $file = basename($reflection->getFileName());
-            $line = $reflection->getStartLine();
-
-            return "Closure in {$file}:{$line}";
-        }
-
-        if (is_object($callback)) {
-            return $callback::class . '::__invoke()';
-        }
-
-        return 'unknown';
     }
 }
